+++ src/components/molecules/vTaskItemStudent.vue
@@ -8,6 +8,9 @@
 				{{ task.courseName }}
 			</v-list-item-subtitle>
 			<v-list-item-title v-text="task.name" />
+			<v-list-item-subtitle>
+				{{ topic }}
+			</v-list-item-subtitle>
 		</v-list-item-content>
 		<v-list-item-action>
 			<v-list-item-action-text
@@ -81,6 +84,12 @@ export default {
 		},
 		defaultIconColor() {
 			return "#54616e";
+		},
+		hasTopic() {
+			return !!this.task.description
+		},
+		topic() {
+			return `${this.$t("pages.tasks.subtitleTopic")} ${this.task.description}`;
 		},
 	},
 	methods: {