+++ CHANGELOG.md
@@ -12,6 +12,12 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Changed
 
+## [26.12.1] - 2021-10-12
+
+### Changed
+
+- BC-379 - hard-code node 14 and npm 6
+
 ## [26.12.0] - 2021-10-12
 
 ### Added